fix: use size_t for size-types
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@2572 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/DeprecatedGUI/CBrowser.cpp
@@ -1256,7 +1256,7 @@ std::string CBrowser::AutoDetectLinks(const std::string text)
 			continue;
 		}
 
-		for (int i = 0; i < sizeof(prefixes)/sizeof(std::string); ++i)  {
+		for (size_t i = 0; i < sizeof(prefixes)/sizeof(std::string); ++i)  {
 			if (text.size() - pos > prefixes[i].size() + 4)  {  // 4 = minimum length of the address, for example a.de
 				if (stringcaseequal(text.substr(pos, prefixes[i].size()), prefixes[i]))  {
